Panic sims

+++ simple_sims/sel_coeff/s_005/simulate_custom.py
@@ -133,11 +133,16 @@ def index_vcf(vcf):
 
 def merge_vcfs(vcf_dir):
     num_files = len(glob(f"{vcf_dir}/*.vcf.sorted.gz"))
-    cmd = f"""bcftools merge -Ov \
-            --force-samples -0 \
-            {" ".join([f"{vcf_dir}/{i}.vcf.sorted.gz" for i in range(num_files)])} > \
-            {vcf_dir}/merged.vcf \
-            """
+    if num_files == 1:
+        cmd = f"""zcat {f"{vcf_dir}/0.vcf.sorted.gz"} > {vcf_dir}/merged.vcf"""
+
+    else:
+        cmd = f"""bcftools merge -Ov -0 \
+                --force-samples --info-rules 'MT:join,S:join' \
+                {" ".join([f"{vcf_dir}/{i}.vcf.sorted.gz" for i in range(num_files)])} > \
+                {vcf_dir}/merged.vcf \
+                """
+
     subprocess.run(cmd, shell=True)
 
 
@@ -150,7 +155,8 @@ def get_num_inds(vcf_file):
 
 def cleanup_intermed(vcf_dir):
     for ifile in glob(f"{vcf_dir}/*"):
-        if "merged" not in ifile:
+        if "merged" not in ifile and "final" not in ifile:
+            pass
             os.remove(ifile)
 
 
@@ -164,6 +170,10 @@ def make_vcf_dir(input_vcf):
             os.remove(ifile)
 
     os.makedirs(vcf_dir, exist_ok=True)
+
+    final_vcf = f"{input_vcf}.final"
+    if os.path.exists(final_vcf):
+        os.rename(final_vcf, f"{vcf_dir}/{final_vcf.split('/')[-1]}")
 
     return vcf_dir
 
@@ -183,13 +193,12 @@ def process_vcfs(input_vcf, num_tps):
             # Now index and merge
             [index_vcf(vcf) for vcf in glob(f"{vcf_dir}/*.vcf")]
             merge_vcfs(vcf_dir)
+
             cleanup_intermed(vcf_dir)
 
     except Exception as e:
         print(f"[ERROR] Couldn't process {e}")
         pass
-
-
 def simulate_prep(
     vcf_file, num_sample_points, slimfile, slim_path, d_block, logfile, dumpFile
 ):
